Split Login function into separate package auth
